+++ include/Ark/Compiler/Macros/Executors/Function.hpp
@@ -2,7 +2,7 @@
  * @file Function.hpp
  * @author Ray John Alovera (rakista112@gmail.com), Alexandre Plateau (lexplt.dev@gmail.com)
  * @brief Executor for List Macros
- * @version 1.1
+ * @version 2.0
  * @date 2021-05-04
  *
  * @copyright Copyright (c) 2021-2024
@@ -26,7 +26,7 @@ namespace Ark::internal
     public:
         using MacroExecutor::MacroExecutor;
 
-        bool applyMacro(Node& node) override;
+        bool applyMacro(Node& node, unsigned depth) override;
         bool canHandle(Node& node) override;
     };