+++ src/arkreactor/Compiler/Macros/Executors/Conditional.cpp
@@ -2,10 +2,10 @@
 
 namespace Ark::internal
 {
-    bool ConditionalExecutor::applyMacro(Node& node)
+    bool ConditionalExecutor::applyMacro(Node& node, unsigned depth)
     {
         Node cond = node.list()[1];
-        const Node temp = evaluate(cond, /* is_not_body */ true);
+        const Node temp = evaluate(cond, depth + 1, /* is_not_body */ true);
         const Node if_true = node.list()[2];
         const Node if_false = node.constList().size() > 3 ? node.list()[3] : getNilNode();