+++ src/arkreactor/Compiler/Macros/Pipeline.cpp
@@ -6,13 +6,13 @@ namespace Ark::internal
         m_executors(executors)
     {}
 
-    bool MacroExecutorPipeline::applyMacro(Node& node) const
+    bool MacroExecutorPipeline::applyMacro(Node& node, unsigned depth) const
     {
         for (const std::shared_ptr<MacroExecutor>& executor : m_executors)
         {
             if (executor->canHandle(node))
             {
-                if (executor->applyMacro(node))
+                if (executor->applyMacro(node, depth))
                     return true;
             }
         }